requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
llama-cpp-python==0.2.11
chromadb==0.4.15
pydantic>=2.6.0